+++ src/feed/index.ts
@@ -3,3 +3,4 @@ export * from './feed.interface';
 export * from './mq-feed';
 export * from './entity-handler.interface';
 export * from './types';
+export * from './utilities';